requirement.txt
Flask
pandas
PyMuPDF
python-docx
spacy
scikit-learn
matplotlib
plotly
reportlab
